+++ src/Options.h
@@ -62,8 +62,8 @@ const std::string NOT_PROVIDED = std::to_string(255);
 // commandline if in fact the option was specified on the commandline, and it will be
 // the default value for the option if the option was not specified on the commandline.
 //
-// To reiterate: by using the OPT_VALUE macro, the value of the option returned
-// will be (in order of priority):
+// To reiterate: by using the OPT_VALUE macro, the value of the option returned will
+// be (in order of priority):
 //
 //    1. the value specified on the grid line IFF the user specified the option on the 
 //       grid line
@@ -100,12 +100,53 @@ const std::string NOT_PROVIDED = std::to_string(255);
 // In that case the value entered by the user on the commandline will be returned IFF
 // the user specified the option on the commandline, otherwise the default value for
 // the option will be returned.
+//
+// Note that the `optName` argument to this macros should be the option name without
+// the "-" or "--" prefix.
 
 
-#define OPT_VALUE(optName, optValue, fallback)  (m_GridLine.optionValues.m_Populated && \
-                                                (!m_GridLine.optionValues.m_VM[optName].defaulted() || !fallback)) \
+#define OPT_VALUE(optName, optValue, fallback)  ((m_GridLine.optionValues.m_Populated && \
+                                                 (!m_GridLine.optionValues.m_VM[optName].defaulted() || !fallback)) \
                                                     ? m_GridLine.optionValues.optValue \
-                                                    : m_CmdLine.optionValues.optValue
+                                                    : m_CmdLine.optionValues.optValue)
+
+
+// OPT_DEFAULTED macro
+//
+// Use to determine, for a given option, whether an option value was specified or it defaulted
+// to the COMPAS default.  Note that this is different from the question of whether the option
+// value is equal to the COMPAS default - this macro indicates how the option value was set:
+// was it provided by the user, or was it set to the default value because a value was not
+// provided by the user.
+//
+// This is a reasonable proxy for the Option::OptionSpecified() function, but only if the
+// `optName` argument is actually a valid option name.
+//
+// The boost defaulted() function will return:
+//
+//     (a) TRUE  if `optName` is a valid option name and a value was not specified by the user
+//         (so Option::OptionSpecified() would return FALSE)
+//     (b) FALSE if `optName` is a valid option name and a value was specified by the user
+//         (so Option::OptionSpecified() would return TRUE)
+//     (c) FALSE if `optName` is not found in the stored list of valid option names
+//         (i.e not a valid option name)
+//
+// In the (a) and (b) cases the boost defaulted() function (and so this macro) is a valid proxy for 
+// Option::OptionSpecified(), but in the (c) case, while the result is technically correct (i.e the 
+// default value was not set for the option), it is not a valid proxy for Option::OptionSpecified()
+// (in this case, Option::OptionSpecified() would return FALSE)
+//
+// When `optName` is known to be a valid option name, this macro is a valid proxy for, and is ~7 times
+// faster than, Option::OptionSpecified() (because it doesn't need to seach the list of valid option
+// names for `optName`): just take the NOT of the OPT_DEFAULTED macro.
+//
+// Note that the `optName` argument to this macros should be the option name without the "-" or "--" prefix.
+
+
+#define OPT_DEFAULTED(optName) (m_GridLine.optionValues.m_Populated \
+                                ? m_GridLine.optionValues.m_VM[optName].defaulted() \
+                                : m_CmdLine.optionValues.m_VM[optName].defaulted())
+
 
 /*
  * Options Singleton
@@ -120,6 +161,81 @@ const std::string NOT_PROVIDED = std::to_string(255);
 class Options {
 
 private:
+
+
+    // The following vectors are used to specify deprecated option strings, option values,
+    // and their replacements (if applicable).
+    //
+    // The vectors below need to be updated whenever we deprecate an option or an option value,
+    // and the option (or value) eventually removed when the deprecation notice period is over.
+    // 
+    //
+    // "deprecatedOptionStrings" vector
+    // --------------------------------
+    //
+    // Each tuple in the "deprecatedOptionStrings" vector records an option that has been deprecated,
+    // but is still available for users to specify.  The tuple entries are:
+    // 
+    //     - the option string for the deprecated option (just the option string - no leading "--")
+    //     - the option string for any replacement for the deprecated option (just the option string,
+    //       no leading "--").  If there is no replacement (i.e. the deprecated option will be removed
+    //       and no replacement option implemented), the replacement option string should be the empty
+    //       string ("")
+    //     - a boolean flag to indicate if the deprecation notice for the option has been shown - should
+    //       be "false" in the vector, and will be set true if and when the deprecation notice for that
+    //       option is shown the first time in a COMPAS run (a deprecation notice for a deprecated option
+    //       is only shown once per COMPAS run).
+    // 
+    //
+    // "deprecatedOptionValues" vector
+    // -------------------------------
+    //
+    // Sometimes we may want to deprecate an option value (e.g. one of the possible mass loss prescriptions).
+    // We may want to do this to rename an option value, or we might want to remove it completely (without
+    // replacement).
+    // 
+    // Each tuple in the "deprecatedOptionValues" vector records an option value that has been deprecated,
+    // but is still available for users to specify.  The tuple entries are:
+    // 
+    //     - the option string for which a value is to be deprecated (just the option string - no leading "--")
+    //     - the value string for the value to be deprecated (e.g. for the value QCRIT_PRESCRIPTION::CLAEYS for
+    //       the option "critical-mass-ratio-prescription", specify "CLAEYS" in the vector)
+    //     - the value string for any replacement value for the deprecated value (e.g. if the value
+    //       QCRIT_PRESCRIPTION::CLAEYS for the option "critical-mass-ratio-prescription", is to be replaced
+    //       with QCRIT_PRESCRIPTION::CLAEYS123, specify "CLAEYS123" in the vector).  If there is no replacement
+    //       (i.e. the deprecated value will be removed and no replacement value implemented), the replacement
+    //       value string should be the empty string ("")
+    //     - a boolean flag to indicate if the deprecation notice for the option valuehas been shown - should
+    //       be "false" in the vector, and will be set true if and when the deprecation notice for that option
+    //       value is shown the first time in a COMPAS run (a deprecation notice for a deprecated option value
+    //       is only shown once per COMPAS run).
+
+    std::vector<std::tuple<std::string, std::string, bool>> deprecatedOptionStrings = {
+        { "black-hole-kicks",                      "black-hole-kicks-mode",                           false },
+        { "chemically-homogeneous-evolution",      "chemically-homogeneous-evolution-mode",           false },
+        { "kick-direction",                        "kick-direction-distribution",                     false },
+        { "luminous-blue-variable-prescription",   "LBV-mass-loss-prescription",                      false },
+        { "mass-transfer",                         "use-mass-transfer",                               false },
+        { "mass-transfer-thermal-limit-accretor",  "mass-transfer-thermal-limit-accretor-multiplier", false },
+        { "OB-mass-loss",                          "OB-mass-loss-prescription",                       false },
+        { "RSG-mass-loss",                         "RSG-mass-loss-prescription",                      false },
+        { "VMS-mass-loss",                         "VMS-mass-loss-prescription",                      false },
+        { "WR-mass-loss",                          "WR-mass-loss-prescription",                       false }
+    };
+
+    std::vector<std::tuple<std::string, std::string, std::string, bool>> deprecatedOptionValues = {
+        { "LBV-mass-loss-prescription",          "NONE", "ZERO", false },
+        { "luminous-blue-variable-prescription", "NONE", "ZERO", false },
+        { "OB-mass-loss",                        "NONE", "ZERO", false },
+        { "OB-mass-loss-prescription",           "NONE", "ZERO", false },
+        { "RSG-mass-loss",                       "NONE", "ZERO", false },
+        { "RSG-mass-loss-prescription",          "NONE", "ZERO", false },
+        { "VMS-mass-loss",                       "NONE", "ZERO", false },
+        { "VMS-mass-loss-prescription",          "NONE", "ZERO", false },
+        { "WR-mass-loss",                        "NONE", "ZERO", false },
+        { "WR-mass-loss-prescription",           "NONE", "ZERO", false }
+    };
+
 
     // The following vectors are used to constrain which options can be specified
     // when:
@@ -213,8 +329,6 @@ private:
 
         "log-level", 
         "log-classes",
-
-        //"logfile-be-binaries",
 
         "logfile-common-envelopes",
         "logfile-common-envelopes-record-types",
@@ -313,8 +427,6 @@ private:
         "allow-touching-at-birth",
         "angular-momentum-conservation-during-circularisation", 
 
-        //"be-binaries",
-
         "case-BB-stability-prescription",
         "circularise-binary-during-mass-transfer",
         "common-envelope-allow-main-sequence-survive",
@@ -377,7 +489,6 @@ private:
         "mass-transfer-accretion-efficiency-prescription",
         "mass-transfer-angular-momentum-loss-prescription",
         "mass-transfer-rejuvenation-prescription",
-        "mass-transfer-thermal-limit-accretor",  // DEPRECATED June 2024 - remove end 2024
         "mass-transfer-thermal-limit-accretor-multiplier",
         "mass-transfer-thermal-limit-C",
         "maximum-mass-donor-nandez-ivanova",
@@ -429,14 +540,10 @@ private:
         "allow-touching-at-birth",
         "angular-momentum-conservation-during-circularisation",
 
-        //"be-binaries",
-
-        "black-hole-kicks", // DEPRECATED June 2024 - remove end 2024
         "black-hole-kicks-mode",
 
         "case-BB-stability-prescription",
         "check-photon-tiring-limit",
-        "chemically-homogeneous-evolution", // DEPRECATED June 2024 - remove end 2024
         "chemically-homogeneous-evolution-mode",
         "circularise-binary-during-mass-transfer",
         "common-envelope-allow-main-sequence-survive",
@@ -477,7 +584,6 @@ private:
         "include-WD-binaries-as-DCO",
         "initial-mass-function", "i",
 
-        "kick-direction",   // DEPRECATED June 2024 - remove end 2024
         "kick-direction-distribution",
         "kick-magnitude-distribution", 
 
@@ -485,9 +591,6 @@ private:
 
         "log-level", 
         "log-classes",
-
-        //"logfile-be-binaries",
-        //"logfile-be-binaries-record-types",
 
         "logfile-common-envelopes",
         "logfile-common-envelopes-record-types",
@@ -507,16 +610,12 @@ private:
         "logfile-system-parameters-record-types",
         "logfile-type",
 
-        "luminous-blue-variable-prescription",  // DEPRECATED June 2024 - remove end 2024
-
         "mass-change-fraction",
         "mass-loss-prescription",
         "mass-ratio-distribution",
-        "mass-transfer",    // DEPRECATED June 2024 - remove end 2024
         "mass-transfer-accretion-efficiency-prescription",
         "mass-transfer-angular-momentum-loss-prescription",
         "mass-transfer-rejuvenation-prescription",
-        "mass-transfer-thermal-limit-accretor", // DEPRECATED June 2024 - remove end 2024
         "mass-transfer-thermal-limit-accretor-multiplier",
         "metallicity-distribution",
         "mode",
@@ -527,7 +626,6 @@ private:
         "neutrino-mass-loss-BH-formation",
         "neutron-star-equation-of-state",
 
-        "OB-mass-loss", // DEPRECATED June 2024 - remove end 2024
         "OB-mass-loss-prescription",
         "orbital-period-distribution",
         "output-container", "c",
@@ -543,7 +641,6 @@ private:
 
         "quiet", 
 
-        "RSG-mass-loss",    // DEPRECATED June 2024 - remove end 2024
         "RSG-mass-loss-prescription",
         "radial-change-fraction",
         "random-seed",
@@ -565,11 +662,9 @@ private:
         "use-mass-loss",
         "use-mass-transfer",
 
-        "VMS-mass-loss",    // DEPRECATED June 2024 - remove end 2024
         "VMW-mass-loss-prescription",
         "version", "v",
 
-        "WR-mass-loss",     // DEPRECATED June 2024 - remove end 2024
         "WR-mass-loss-prescription",
 
         "yaml-template"
@@ -604,9 +699,6 @@ private:
 
         "log-classes",
         "log-level", 
-
-        //"logfile-be-binaries",
-        //"logfile-be-binaries-record-types",
 
         "logfile-common-envelopes",
         "logfile-common-envelopes-record-types",
@@ -1089,8 +1181,6 @@ public:
                 vm[opt].value() = boost::any(val);
             }
 
-            int         OptionSpecified(std::string p_OptionString);
-
             std::string SetCalculatedOptionDefaults(const BOOST_MAP p_BoostMap);
 
         public:
@@ -1213,12 +1303,15 @@ public:
 
     int                         ApplyNextGridLine();
 
+    std::string                 CheckDeprecatedOptionString(const std::string p_OptionString);
+    std::string                 CheckDeprecatedOptionValue(const std::string p_OptionString, const std::string p_OptionValue);
     void                        CloseGridFile() { m_Gridfile.handle.close(); m_Gridfile.filename = ""; m_Gridfile.error = ERROR::EMPTY_FILENAME; }
 
     bool                        Initialise(int p_OptionCount, char *p_OptionStrings[]);
     bool                        InitialiseEvolvingObject(const std::string p_OptionsString);
 
     ERROR                       OpenGridFile(const std::string p_GridFilename);
+    bool                        OptionDefaulted(const std::string p_OptionString) const                                 { return OPT_DEFAULTED(p_OptionString); }
     bool                        OptionSpecified(const std::string p_OptionString);
 
     COMPAS_VARIABLE             OptionValue(const T_ANY_PROPERTY p_Property) const;
@@ -1230,8 +1323,6 @@ public:
     ERROR                       SeekToGridFileLine(const unsigned int p_Line);
 
     std::string                 SetRandomSeed(const unsigned long int p_RandomSeed, const OPTIONS_ORIGIN p_OptionsSet);
-
-    void                        ShowDeprecations(const bool p_Commandline = true);
 
     // getters
 
@@ -1246,13 +1337,13 @@ public:
     bool                                        AngularMomentumConservationDuringCircularisation() const                { return OPT_VALUE("angular-momentum-conservation-during-circularisation", m_AngularMomentumConservationDuringCircularisation, true); }
 
 
-    BLACK_HOLE_KICKS_MODE                       BlackHoleKicksMode() const                                              { return OPTIONS->OptionSpecified("black-hole-kicks-mode") ? OPT_VALUE("black-hole-kicks-mode", m_BlackHoleKicksMode.type, true) : OPT_VALUE("black-hole-kicks", m_BlackHoleKicksMode.type, true); } // black-hole-kicks DEPRECATED June 2024 - remove end 2024
+    BLACK_HOLE_KICKS_MODE                       BlackHoleKicksMode() const                                              { return OPT_VALUE("black-hole-kicks-mode", m_BlackHoleKicksMode.type, true); }
     
     CASE_BB_STABILITY_PRESCRIPTION              CaseBBStabilityPrescription() const                                     { return OPT_VALUE("case-BB-stability-prescription", m_CaseBBStabilityPrescription.type, true); }
     
     bool                                        CheckPhotonTiringLimit() const                                          { return OPT_VALUE("check-photon-tiring-limit", m_CheckPhotonTiringLimit, true); }
 
-    CHE_MODE                                    CHEMode() const                                                         { return OPTIONS->OptionSpecified("chemically-homogeneous-evolution-mode") ? OPT_VALUE("chemically-homogeneous-evolution-mode", m_CheMode.type, true) : OPT_VALUE("chemically-homogeneous-evolution", m_CheMode.type, true); } // chemically-homogeneous-evolution DEPRECATED June 2024 - remove end 2024
+    CHE_MODE                                    CHEMode() const                                                         { return OPT_VALUE("chemically-homogeneous-evolution-mode", m_CheMode.type, true); }
 
     bool                                        CirculariseBinaryDuringMassTransfer() const                             { return OPT_VALUE("circularise-binary-during-mass-transfer", m_CirculariseBinaryDuringMassTransfer, true); }
 
@@ -1331,7 +1422,7 @@ public:
     double                                      InitialMassFunctionMin() const                                          { return OPT_VALUE("initial-mass-min", m_InitialMassFunctionMin, true); }
     double                                      InitialMassFunctionPower() const                                        { return OPT_VALUE("initial-mass-power", m_InitialMassFunctionPower, true); }
 
-    KICK_DIRECTION_DISTRIBUTION                 KickDirectionDistribution() const                                       { return OPTIONS->OptionSpecified("kick-direction-distribution") ? OPT_VALUE("kick-direction-distribution", m_KickDirectionDistribution.type, true) : OPT_VALUE("kick-direction", m_KickDirectionDistribution.type, true); } // kick-direction DEPRECATED June 2024 - remove end 2024
+    KICK_DIRECTION_DISTRIBUTION                 KickDirectionDistribution() const                                       { return OPT_VALUE("kick-direction-distribution", m_KickDirectionDistribution.type, true); }
     double                                      KickDirectionPower() const                                              { return OPT_VALUE("kick-direction-power", m_KickDirectionPower, true); }
     double                                      KickScalingFactor() const                                               { return OPT_VALUE("kick-scaling-factor", m_KickScalingFactor, true); }
     KICK_MAGNITUDE_DISTRIBUTION                 KickMagnitudeDistribution() const                                       { return OPT_VALUE("kick-magnitude-distribution", m_KickMagnitudeDistribution.type, true); }
@@ -1400,7 +1491,7 @@ public:
     double                                      LuminosityToMassThreshold() const                                       { return OPT_VALUE("luminosity-to-mass-threshold", m_LuminosityToMassThreshold, true); }
 
     double                                      LuminousBlueVariableFactor() const                                      { return OPT_VALUE("luminous-blue-variable-multiplier", m_LuminousBlueVariableFactor, true); }
-    LBV_MASS_LOSS_PRESCRIPTION                  LBVMassLossPrescription() const                                         { return OPTIONS->OptionSpecified("LBV-mass-loss-prescription") ? OPT_VALUE("LBV-mass-loss-prescription", m_LBVMassLossPrescription.type, true) : OPT_VALUE("luminous-blue-variable-prescription", m_LBVMassLossPrescription.type, true); } // luminous-blue-variable-prescription DEPRECATED June 2024 - remove end 2024
+    LBV_MASS_LOSS_PRESCRIPTION                  LBVMassLossPrescription() const                                         { return OPT_VALUE("LBV-mass-loss-prescription", m_LBVMassLossPrescription.type, true); }
     
     double                                      MassChangeFraction() const                                              { return m_CmdLine.optionValues.m_MassChangeFraction; }
     
@@ -1437,7 +1528,7 @@ public:
     double                                      MassTransferJlossMacLeodLinearFractionDegen() const                     { return OPT_VALUE("mass-transfer-jloss-macleod-linear-fraction-degen", m_MassTransferJlossMacLeodLinearFractionDegen, true); }
     double                                      MassTransferJlossMacLeodLinearFractionNonDegen() const                  { return OPT_VALUE("mass-transfer-jloss-macleod-linear-fraction-non-degen", m_MassTransferJlossMacLeodLinearFractionNonDegen, true); }
     MT_REJUVENATION_PRESCRIPTION                MassTransferRejuvenationPrescription() const                            { return OPT_VALUE("mass-transfer-rejuvenation-prescription", m_MassTransferRejuvenationPrescription.type, true); }
-    MT_THERMALLY_LIMITED_VARIATION              MassTransferThermallyLimitedVariation() const                           { return OPTIONS->OptionSpecified("mass-transfer-thermal-limit-accretor-multiplier") ? OPT_VALUE("mass-transfer-thermal-limit-accretor-multiplier", m_MassTransferThermallyLimitedVariation.type, true) : OPT_VALUE("mass-transfer-thermal-limit-accretor", m_MassTransferThermallyLimitedVariation.type, true); } // mass-transfer-thermal-limit-accretor DEPRECATED June 2024 - remove end 2024
+    MT_THERMALLY_LIMITED_VARIATION              MassTransferThermallyLimitedVariation() const                           { return OPT_VALUE("mass-transfer-thermal-limit-accretor-multiplier", m_MassTransferThermallyLimitedVariation.type, true); }
     double                                      MaxEvolutionTime() const                                                { return OPT_VALUE("maximum-evolution-time", m_MaxEvolutionTime, true); }
     double                                      MaximumNeutronStarMass() const                                          { return OPT_VALUE("maximum-neutron-star-mass", m_MaximumNeutronStarMass, true); }
     unsigned long int                           MaxNumberOfTimestepIterations() const                                   { return OPT_VALUE("maximum-number-timestep-iterations", m_MaxNumberOfTimestepIterations, true); }
@@ -1467,7 +1558,7 @@ public:
     std::vector<std::string>                    NotesHdrs() const                                                       { return m_CmdLine.optionValues.m_NotesHdrs; }
  
     size_t                                      nObjectsToEvolve() const                                                { return m_CmdLine.optionValues.m_ObjectsToEvolve; }
-    OB_MASS_LOSS_PRESCRIPTION                   OBMassLossPrescription() const                                          { return OPTIONS->OptionSpecified("OB-mass-loss-prescription") ? OPT_VALUE("OB-mass-loss-prescription", m_OBMassLossPrescription.type, true) : OPT_VALUE("OB-mass-loss", m_OBMassLossPrescription.type, true); } // OB-mass-loss DEPRECATED June 2024 - remove end 2024
+    OB_MASS_LOSS_PRESCRIPTION                   OBMassLossPrescription() const                                          { return OPT_VALUE("OB-mass-loss-prescription", m_OBMassLossPrescription.type, true); }
     bool                                        OptimisticCHE() const                                                   { return CHEMode() == CHE_MODE::OPTIMISTIC; }
 
     double                                      OrbitalPeriod() const                                                   { return OPT_VALUE("orbital-period", m_OrbitalPeriod, true); }
@@ -1534,7 +1625,7 @@ public:
     double                                      RotationalFrequency() const                                             { return OPT_VALUE("rotational-frequency", m_RotationalFrequency, true); }
     double                                      RotationalFrequency1() const                                            { return OPT_VALUE("rotational-frequency-1", m_RotationalFrequency1, true); }
     double                                      RotationalFrequency2() const                                            { return OPT_VALUE("rotational-frequency-2", m_RotationalFrequency2, true); }
-    RSG_MASS_LOSS_PRESCRIPTION                  RSGMassLossPrescription() const                                         { return OPTIONS->OptionSpecified("RSG-mass-loss-prescription") ? OPT_VALUE("RSG-mass-loss-prescription", m_RSGMassLossPrescription.type, true) : OPT_VALUE("RSG-mass-loss", m_RSGMassLossPrescription.type, true); } // RSG-mass-loss DEPRECATED June 2024 - remove end 2024
+    RSG_MASS_LOSS_PRESCRIPTION                  RSGMassLossPrescription() const                                         { return OPT_VALUE("RSG-mass-loss-prescription", m_RSGMassLossPrescription.type, true); }
 
     bool                                        ScaleCHEMassLossWithSurfaceHeliumAbundance() const                      { return OPT_VALUE("scale-CHE-mass-loss-with-surface-helium-abundance", m_ScaleCHEMassLossWithSurfaceHeliumAbundance, false); }
     double                                      ScaleTerminalWindVelocityWithMetallicityPower() const                   { return OPT_VALUE("scale-terminal-wind-velocity-with-metallicity-power", m_ScaleTerminalWindVelocityWithMetallicityPower, true);}
@@ -1564,13 +1655,13 @@ public:
 
     bool                                        UseFixedUK() const                                                      { return (m_GridLine.optionValues.m_UseFixedUK || m_CmdLine.optionValues.m_UseFixedUK); }
     bool                                        UseMassLoss() const                                                     { return OPT_VALUE("use-mass-loss", m_UseMassLoss, true); }
-    bool                                        UseMassTransfer() const                                                 { return OPTIONS->OptionSpecified("use-mass-transfer") ? OPT_VALUE("use-mass-transfer", m_UseMassTransfer, true) : OPT_VALUE("mass-transfer", m_UseMassTransfer, true); } // mass-loss DEPRECATED June 2024 - remove end 2024
+    bool                                        UseMassTransfer() const                                                 { return OPT_VALUE("use-mass-transfer", m_UseMassTransfer, true); }
     bool                                        UsePairInstabilitySupernovae() const                                    { return OPT_VALUE("pair-instability-supernovae", m_UsePairInstabilitySupernovae, true); }
     bool                                        UsePulsationalPairInstability() const                                   { return OPT_VALUE("pulsational-pair-instability", m_UsePulsationalPairInstability, true); }
 
-    VMS_MASS_LOSS_PRESCRIPTION                  VMSMassLossPrescription() const                                         { return OPTIONS->OptionSpecified("VMS-mass-loss-prescription") ? OPT_VALUE("VMS-mass-loss-prescription", m_VMSMassLossPrescription.type, true) : OPT_VALUE("VMS-mass-loss", m_VMSMassLossPrescription.type, true); } // VMS-mass-loss DEPRECATED June 2024 - remove end 2024
+    VMS_MASS_LOSS_PRESCRIPTION                  VMSMassLossPrescription() const                                         { return OPT_VALUE("VMS-mass-loss-prescription", m_VMSMassLossPrescription.type, true); }
     double                                      WolfRayetFactor() const                                                 { return OPT_VALUE("wolf-rayet-multiplier", m_WolfRayetFactor, true); }
-    WR_MASS_LOSS_PRESCRIPTION                   WRMassLossPrescription() const                                          { return OPTIONS->OptionSpecified("WR-mass-loss-prescription") ? OPT_VALUE("WR-mass-loss-prescription", m_WRMassLossPrescription.type, true) : OPT_VALUE("WR-mass-loss", m_WRMassLossPrescription.type, true); } // WR-mass-loss DEPRECATED June 2024 - remove end 2024
+    WR_MASS_LOSS_PRESCRIPTION                   WRMassLossPrescription() const                                          { return OPT_VALUE("WR-mass-loss-prescription", m_WRMassLossPrescription.type, true); }
     std::string                                 YAMLfilename() const                                                    { return m_CmdLine.optionValues.m_YAMLfilename; }
     std::string                                 YAMLtemplate() const                                                    { return m_CmdLine.optionValues.m_YAMLtemplate; }